Fixed broken shakespeare_with_tpu_and_keras.ipynb example
Not sure why this example is now broken on Colab.
Thinking version control problem between numpy and tensorflow.

In the current version of the example, prediction breaks on the second prediction because it is expecting an array of single ascii value arrays (like the first seeded input).

However the array that is appended for the next prediction is just an array of ascii values, which causes the predict function to throw an error.

Also, the printing section is expecting an array of ascii values instead of the array of  single value ascii arrays used for storing predictions.

The fixes are minor, simply making sure each character is  created as a single ascii value array, and then indexed that way during printing of results.

Side note: I got much better results by increasing (4X)  the number of epochs, step size and embedding dimensions during training. Results in a much higher resolution weights file that produces much more cogent outputs.